combined into one entity "terminal"
